KmgFundApplicationのコメントを修正し、KmgFundGenMessageTypes、KmgFundLogMessageTypes、KmgFundValMessageTypesを新規作成しました。また、KmgYamlUtilsでのインポート文を修正しました。
