+++ src/main/java/kmg/fund/domain/types/msg/KmgFundGenMessageTypes.java
@@ -1,12 +1,16 @@
-package kmg.fund.domain.types;
+package kmg.fund.domain.types.msg;
 
 import java.util.HashMap;
 import java.util.Map;
 
-import kmg.fund.infrastructure.common.KmgFundComLogMessageTypes;
+import kmg.fund.infrastructure.common.KmgFundComExcMessageTypes;
+import kmg.fund.infrastructure.common.KmgFundComGenMessageTypes;
 
 /**
- * KMG 基盤ログメッセージの種類<br>
+ * KMG 基盤一般メッセージの種類<br>
+ * <p>
+ * Msgは、Messageの略。
+ * </p>
  *
  * @author KenichiroArai
  *
@@ -14,7 +18,7 @@ import kmg.fund.infrastructure.common.KmgFundComLogMessageTypes;
  *
  * @version 0.1.0
  */
-public enum KmgFundLogMessageTypes implements KmgFundComLogMessageTypes {
+public enum KmgFundGenMessageTypes implements KmgFundComGenMessageTypes, KmgFundComExcMessageTypes {
 
     /* 定義：開始 */
 
@@ -29,7 +33,30 @@ public enum KmgFundLogMessageTypes implements KmgFundComLogMessageTypes {
      */
     NONE("指定無し"),
 
+    /**
+     * 該当するYAMLファイルがありません。ファイルパス=[{0}]
+     *
+     * @author KenichiroArai
+     *
+     * @since 0.1.0
+     *
+     * @version 0.1.0
+     */
+    KMGFUND_GEN24000("該当するYAMLファイルがありません。ファイルパス=[{0}]"),
+
+    /**
+     * YAMLにロードするファイルの読み込みに失敗しました。ファイルパス=[{0}]
+     *
+     * @author KenichiroArai
+     *
+     * @since 0.1.0
+     *
+     * @version 0.1.0
+     */
+    KMGFUND_GEN24001("YAMLにロードするファイルの読み込みに失敗しました。ファイルパス=[{0}]"),
+
     /* 定義：終了 */
+
     ;
 
     /**
@@ -37,14 +64,14 @@ public enum KmgFundLogMessageTypes implements KmgFundComLogMessageTypes {
      *
      * @since 0.1.0
      */
-    private static final Map<String, KmgFundLogMessageTypes> VALUES_MAP = new HashMap<>();
+    private static final Map<String, KmgFundGenMessageTypes> VALUES_MAP = new HashMap<>();
 
     static {
 
         /* 種類のマップにプット */
-        for (final KmgFundLogMessageTypes type : KmgFundLogMessageTypes.values()) {
+        for (final KmgFundGenMessageTypes type : KmgFundGenMessageTypes.values()) {
 
-            KmgFundLogMessageTypes.VALUES_MAP.put(type.get(), type);
+            KmgFundGenMessageTypes.VALUES_MAP.put(type.get(), type);
 
         }
 
@@ -85,9 +112,9 @@ public enum KmgFundLogMessageTypes implements KmgFundComLogMessageTypes {
      *
      * @return デフォルト値
      */
-    public static KmgFundLogMessageTypes getDefault() {
+    public static KmgFundGenMessageTypes getDefault() {
 
-        final KmgFundLogMessageTypes result = NONE;
+        final KmgFundGenMessageTypes result = NONE;
         return result;
 
     }
@@ -105,9 +132,9 @@ public enum KmgFundLogMessageTypes implements KmgFundComLogMessageTypes {
      *
      * @return 種類。指定無し（NONE）：キーが存在しない場合。
      */
-    public static KmgFundLogMessageTypes getEnum(final String key) {
+    public static KmgFundGenMessageTypes getEnum(final String key) {
 
-        KmgFundLogMessageTypes result = KmgFundLogMessageTypes.VALUES_MAP.get(key);
+        KmgFundGenMessageTypes result = KmgFundGenMessageTypes.VALUES_MAP.get(key);
 
         if (result == null) {
 
@@ -125,9 +152,9 @@ public enum KmgFundLogMessageTypes implements KmgFundComLogMessageTypes {
      *
      * @return 初期値
      */
-    public static KmgFundLogMessageTypes getInitValue() {
+    public static KmgFundGenMessageTypes getInitValue() {
 
-        final KmgFundLogMessageTypes result = NONE;
+        final KmgFundGenMessageTypes result = NONE;
         return result;
 
     }
@@ -140,7 +167,7 @@ public enum KmgFundLogMessageTypes implements KmgFundComLogMessageTypes {
      * @param displayName
      *                    表示名
      */
-    KmgFundLogMessageTypes(final String displayName) {
+    KmgFundGenMessageTypes(final String displayName) {
 
         this.displayName = displayName;
         this.key = super.name();
